+++ packages/frontend/pages/research.tsx
@@ -1,212 +1,145 @@
 import React from 'react'
 import Link from 'next/link'
-import Image from 'next/image'
 import { Hidden, Button, Typography, useMediaQuery, useTheme, Paper, Collapse } from '@material-ui/core'
 import { createStyles, makeStyles } from '@material-ui/core/styles'
-import Hamburger from 'hamburger-react'
 
-import logo from 'public/images/OpynLogo.svg'
+import AthenaBackgroundImg from 'public/images/landing/athena1-desktop.png'
 import DefaultSiteSeo from '@components/DefaultSiteSeo/DefaultSiteSeo'
+import { Nav } from '@components/Nav/Basic'
 
 const useStyles = makeStyles((theme) =>
   createStyles({
-    m_nav: {
-      padding: `0px 10px`,
-      display: 'flex',
-      alignItems: 'center',
-      boxShadow: '0px 3px 4px rgba(0, 0, 0, 0.2)',
-      background: 'linear-gradient(180deg, #1A1C1D 0%, #191B1C 100%)',
-    },
-    m_navMenu: {
-      margin: `0px 4px 0px -4px`,
+    athenaBackground: {
       position: 'absolute',
-    },
-    m_navLogo: { flex: 1, alignItems: 'center', justifyContent: 'center', display: 'flex' },
-    m_navDrawer: {},
-    m_navLinks: {
-      display: 'flex',
-      flexDirection: 'column',
-      gap: `15px`,
-      flex: 1,
-    },
-    m_drawer: {
-      position: 'absolute',
+      backgroundImage: `url(${AthenaBackgroundImg.src})`,
+      height: `calc(100vh - 120px)`,
       width: '100vw',
-      zIndex: 1,
+      backgroundRepeat: 'no-repeat',
+      backgroundPosition: 'right',
+      backgroundSize: 'contain',
+      zIndex: -1,
     },
-    m_drawerWrapper: {
-      backgroundColor: '#232526',
-      padding: '20px 20px',
+    content: {
+      width: '85%',
+      margin: '3em auto 3em',
+      paddingBottom: '1em',
     },
-    m_navLink: {
-      fontFamily: 'DM Sans',
+    heading: {
+      fontSize: '32px',
       fontWeight: 700,
-      fontSize: `16px`,
-      lineHeight: '140%',
-      color: '#F2F2F2',
-      opacity: 0.5,
-      cursor: 'pointer',
-      transitionDuration: '300ms',
-      '&:hover': {
-        opacity: 1,
-      },
-    },
-    m_navAction: {
-      position: 'absolute',
-      right: '10px',
-    },
-    m_navStartEarningButton: {
-      backgroundColor: theme.palette.primary.main,
-      padding: `8px 10px`,
-      fontFamily: 'DM Sans',
-      fontWeight: 800,
-      fontSize: '14px',
+      letterSpacing: '-0.02em',
       lineHeight: '130%',
-      maxWidth: '250px',
-      '&:hover': {
-        backgroundColor: theme.palette.primary.dark,
-      },
     },
-    logo: {
-      cursor: 'pointer',
-    },
-    nav: {
-      padding: '8px 120px',
-      borderBottom: '1px solid #333333',
-      display: 'flex',
-      alignItems: 'center',
-    },
-    navLinks: {
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'flex-end',
-      gap: '32px',
-      flex: 1,
-    },
-    navLink: {
-      fontFamily: 'DM Sans',
-      fontWeight: 700,
-      fontSize: `16px`,
-      lineHeight: '140%',
-      color: '#F2F2F2',
-      opacity: 0.5,
-      cursor: 'pointer',
-      transitionDuration: '300ms',
-      '&:hover': {
-        opacity: 1,
-      },
-    },
-    navAction: {
-      marginLeft: '16px',
-    },
-    navStartEarningButton: {
-      backgroundColor: theme.palette.primary.main,
-      padding: '12px 16px',
-      fontFamily: 'DM Sans',
-      fontWeight: 700,
+
+    ul: {
+      listStyleType: 'upper-roman',
       fontSize: '16px',
-      lineHeight: '130%',
-      maxWidth: '180px',
-      '&:hover': {
-        backgroundColor: theme.palette.primary.dark,
+      padding: 0,
+      paddingLeft: '1em',
+      margin: 0,
+      marginTop: '2em',
+      color: theme.palette.text.hint,
+      '& li': {
+        marginBottom: '1.5em',
+        fontSize: '20px',
+      },
+      [theme.breakpoints.up('md')]: {
+        '& li': {
+          marginBottom: '2em',
+        },
       },
     },
-    navMenu: {
-      margin: `0 10px`,
-      position: 'absolute',
+    liContainer: {
+      paddingLeft: 4,
     },
-    drawerWrapper: {
-      backgroundColor: '#232526',
-      padding: '20px 20px',
+    link: {
+      display: 'inline-block',
+      color: theme.palette.primary.main,
+      fontWeight: 500,
+      fontSize: '18px',
+      fontFamily: 'DM Sans',
+      marginTop: theme.spacing(0.5),
+
+      '&:hover': {
+        color: theme.palette.primary.dark,
+        textDecoration: 'underline',
+      },
+      [theme.breakpoints.up('md')]: {
+        fontSize: '24px',
+      },
+    },
+    subtext: {
+      fontSize: '14px',
+      fontWeight: 500,
+      fontFamily: 'DM Mono',
+      color: theme.palette.text.hint,
+      maxWidth: '80%',
+      [theme.breakpoints.up('md')]: {
+        fontSize: '16px',
+        maxWidth: '600px',
+      },
+    },
+    authors: {
+      fontWeight: 700,
     },
   }),
 )
 
-const navLinks = [
-  { label: 'Strategies', link: '/strategies/crab' },
-  { label: 'Squeeth', link: '/squeeth' },
-  { label: 'Auction', link: 'https://squeethportal.xyz' },
+const RESEARCH_PAPERS = [
   {
-    label: 'FAQ',
-    link: 'https://opyn.gitbook.io/opyn-strategies/strategies-faq/faq',
+    id: 1,
+    title: 'Power Perpetuals',
+    link: 'https://www.paradigm.xyz/2021/08/power-perpetuals',
+    authors: ['Dave White', 'Dan Robinson', 'Zubin Koticha', 'Andrew Leone', 'Alexis Gauba', 'Aparna Krishnan'],
+    date: 'Aug 17, 2021',
   },
+  {
+    id: 2,
+    title: 'Squeeth Mechanism',
+    link: 'https://medium.com/opyn/squeeth-insides-volume-1-funding-and-volatility-f16bed146b7d',
+    authors: ['Joseph Clark'],
+    date: 'Dec 21, 2021',
+  },
+  {
+    id: 3,
+    title: 'Squeeth Primer',
+    link: 'https://medium.com/opyn/squeeth-primer-a-guide-to-understanding-opyns-implementation-of-squeeth-a0f5e8b95684',
+    authors: ['Wade Prospere'],
+    date: 'Jan 10, 2022',
+  },
+  // add more papers as needed
 ]
 
-function Nav() {
-  const classes = useStyles()
-  const theme = useTheme()
-  const isMobileBreakpoint = useMediaQuery(theme.breakpoints.down('sm'))
-
-  const [navOpen, setNavOpen] = React.useState(false)
-
-  if (isMobileBreakpoint) {
-    return (
-      <div>
-        <div className={classes.m_nav}>
-          <div className={classes.m_navMenu}>
-            <Hamburger size={20} toggled={navOpen} toggle={setNavOpen} />
-          </div>
-          <div className={classes.m_navLogo}>
-            <Link href={'/'} passHref>
-              <Image src={logo} alt="logo" width={83} height={59} />
-            </Link>
-          </div>
-          <div className={classes.m_navAction}>
-            <Link href={'/strategies/crab'} passHref>
-              <Button className={classes.m_navStartEarningButton}>Launch</Button>
-            </Link>
-          </div>
-        </div>
-
-        <div style={{ position: 'absolute', width: '100vw', zIndex: 1 }}>
-          <Collapse in={navOpen}>
-            <Paper className={classes.m_drawerWrapper} elevation={2}>
-              <div className={classes.m_navLinks}>
-                {navLinks.map((link) => (
-                  <Typography variant="h3" className={classes.m_navLink} key={link.label}>
-                    <Link href={link.link} passHref>
-                      {link.label}
-                    </Link>
-                  </Typography>
-                ))}
-              </div>
-            </Paper>
-          </Collapse>
-        </div>
-      </div>
-    )
-  }
-
-  return (
-    <div className={classes.nav}>
-      <div className={classes.logo}>
-        <Link href={'/'} passHref>
-          <Image src={logo} alt="logo" width={97} height={75} />
-        </Link>
-      </div>
-      <div className={classes.navLinks}>
-        {navLinks.map((link) => (
-          <Typography variant="h3" className={classes.navLink} key={link.label}>
-            <Link href={link.link} passHref>
-              {link.label}
-            </Link>
-          </Typography>
-        ))}
-        <div className={classes.navAction}>
-          <Link href={'/strategies/crab'} passHref>
-            <Button className={classes.navStartEarningButton}>Launch App</Button>
-          </Link>
-        </div>
-      </div>
-    </div>
-  )
-}
-
 function ResearchPage() {
+  const classes = useStyles()
+
   return (
     <div>
       <DefaultSiteSeo />
       <Nav />
+
+      <div className={classes.athenaBackground} />
+
+      <div className={classes.content}>
+        <Typography variant="h3" className={classes.heading}>
+          Research
+        </Typography>
+
+        <ul className={classes.ul}>
+          {RESEARCH_PAPERS.map((paper) => (
+            <li key={paper.id}>
+              <div className={classes.liContainer}>
+                <a href={paper.link} target="_blank" rel="noopener noreferrer">
+                  <Typography className={classes.link}>{paper.title}</Typography>
+                </a>
+                <Typography className={classes.subtext}>
+                  by <span className={classes.authors}>{paper.authors.join(', ')}</span> on <b>{paper.date}</b>
+                </Typography>
+              </div>
+            </li>
+          ))}
+        </ul>
+      </div>
     </div>
   )
 }
